+++ tests/wildboar/transform/test_transform.py
@@ -1,13 +1,16 @@
 # Authors: Isak Samsten
 # License: BSD 3 clause
+import numpy as np
 import pytest
 from numpy.testing import assert_almost_equal
 from sklearn.utils.estimator_checks import check_estimators_pickle
+
 from wildboar.datasets import load_gun_point
 from wildboar.transform import (
     PAA,
     SAX,
     CastorTransform,
+    DerivativeTransform,
     DilatedShapeletTransform,
     HydraTransform,
     IntervalTransform,
@@ -55,3 +58,10 @@ def test_random_shapelet_transform_equals_shapelet_transform_strategy_random():
         RandomShapeletTransform(random_state=1).fit_transform(X),
         ShapeletTransform(strategy="random", random_state=1).fit_transform(X),
     )
+
+
+def test_derivative_transform():
+    t = DerivativeTransform(method="forward")
+    X = np.array([[1, 7, 3, 4], [4, 2, 6, 7]])
+    expected_output = np.array([[3.5, 3.5, -2.75, -2.75], [-0.5, -0.5, 3.25, 3.25]])
+    assert_almost_equal(t.fit_transform(X), expected_output)